# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Translation API",
        "description": "API documentation for the translation management system.",
        "contact": {
            "email": "dev@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Dev Server"
        }
    ],
    "paths": {
        "/api/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "List all translations",
                "operationId": "03e0745fb3bed90fe04f46a1395fa4c1",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create a new translation",
                "operationId": "711dfe2c7f0dbd38f67d0f3e7e07f5b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "locale",
                                    "key",
                                    "content"
                                ],
                                "properties": {
                                    "locale": {
                                        "type": "string"
                                    },
                                    "key": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/translations/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export all translations grouped by locale",
                "operationId": "9f227c61c766f0f5cacdb54bab6df4a2",
                "responses": {
                    "200": {
                        "description": "Exported JSON"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/translations/search": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Search translations by key, content, or tag",
                "operationId": "0e8e57c1d0fe6abb8a7320538657df42",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Translations",
            "description": "Translations"
        }
    ]
}